# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: GreenRoot API
  version: 0.1.0
  description: Modular Go chi backend for the GreenRoot platform.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user profile
    put:
      tags: [Users]
      summary: Update current user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user profile
...
  /api/v1/plant-requests:
    get:
      tags: [Plant Requests]
      summary: List plant requests
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: nursery_id
          in: query
          schema:
            type: integer
        - name: plant_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Plant request list
    post:
      tags: [Plant Requests]
      summary: Create plant request
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantRequestRequest"
      responses:
        "201":
          description: Plant request created
...
  /api/v1/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
            format: int64
        - name: subscription_status
          in: query
          schema:
            type: string
            enum: [ACTIVE, PAUSED, CANCELLED, EXPIRED]
      responses:
        "200":
          description: Subscription list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionsResponse"
    post:
      tags: [Subscriptions]
      summary: Create user subscription
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionCreateRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionResponse"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateProfileRequest:
      type: object
      required: [first_name]
      properties:
        first_name:
          type: string
          example: Demo
        last_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        profile_image_url:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
          enum: [MALE, FEMALE, NON_BINARY, OTHER, PREFER_NOT_TO_SAY]
...
    PlantRequestRequest:
      type: object
      required: [requesting_nursery_id, plant_id, quantity_required]
      properties:
        requesting_nursery_id:
          type: integer
          format: int64
        plant_id:
          type: integer
          format: int64
        size_id:
          type: integer
          nullable: true
        quantity_required:
          type: integer
          minimum: 1
        radius_km:
          type: integer
          minimum: 1
          example: 50
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [OPEN, RESPONDED, FULFILLED, CANCELLED, EXPIRED]
        expires_at:
          type: string
          format: date-time
          nullable: true
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          format: int64
        subscription_code:
          type: string
          example: SUB-000001
        user_id:
          type: integer
          format: int64
        plan_id:
          type: integer
          format: int64
        plan_code:
          type: string
        plan_name:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
          nullable: true
        subscription_status:
          type: string
          enum: [ACTIVE, PAUSED, CANCELLED, EXPIRED]
        auto_renew:
          type: boolean
        latest_payment:
          $ref: "#/components/schemas/SubscriptionPaymentSummary"
    SubscriptionPaymentSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        payment_code:
          type: string
          example: PAY-20260622-0001
        amount:
          type: number
        payment_method:
          type: string
          nullable: true
        transaction_reference:
          type: string
          nullable: true
        payment_status:
          type: string
        provider:
          type: string
          nullable: true
        provider_payment_id:
          type: string
          nullable: true
        provider_order_id:
          type: string
          nullable: true
...
    SubscriptionsResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        pagination:
          type: object
    SubscriptionResponse:
      type: object
      properties:
        subscription:
          $ref: "#/components/schemas/Subscription"
    SubscriptionCreateRequest:
      type: object
      required: [plan_id]
      properties:
        user_id:
          type: integer
          format: int64
          nullable: true
        plan_id:
          type: integer
          format: int64
        billing_cycle:
          type: string
          enum: [MONTHLY, YEARLY]
          default: MONTHLY
        start_date:
          type: string
          format: date
          nullable: true
        auto_renew:
          type: boolean
        payment_method:
          type: string
          enum: [UPI, CARD, CASH, BANK_TRANSFER, NET_BANKING, WALLET, COD, CHEQUE, OTHER]
          nullable: true
        provider:
          type: string
          nullable: true
          example: RAZORPAY
        provider_order_id:
          type: string
          nullable: true
...